services/radiomics-calculator/pyproject.toml
[project]
name = "radiomics-calculator"
version = "0.1.0"
description = "Radiomics feature extraction pipeline"
requires-python = ">=3.12"
dependencies = [
    "imaging-common",
    "numpy",
    "pyyaml",
    "croniter",
    "psycopg2-binary>=2.9.6",
    "requests",
    "pydicom<3",
]

[project.optional-dependencies]
runtime = [
    "pyradiomics",
    "simpleitk",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/radiomics_calculator"]

[tool.uv.sources]
imaging-common = { workspace = true }
pyradiomics = { git = "https://github.com/AIM-Harvard/pyradiomics.git" }
